refactor: split internal locations in separate file
