+++ packages/lib/event-types/transformers/api-to-internal/locations.ts
@@ -1,6 +1,14 @@
-import { z } from "zod";
-
 import { type CreateEventTypeInput_2024_06_14, type Integration_2024_06_14 } from "@calcom/platform-types";
+
+import type {
+  AttendeeAddressLocation,
+  AttendeeDefinedLocation,
+  AttendeePhoneLocation,
+  OrganizerAddressLocation,
+  OrganizerIntegrationLocation,
+  OrganizerLinkLocation,
+  OrganizerPhoneLocation,
+} from "../internal/locations";
 
 const integrationsMapping: Record<Integration_2024_06_14, "integrations:daily"> = {
   "cal-video": "integrations:daily",
@@ -48,65 +56,3 @@ export function transformLocationsApiToInternal(
     }
   });
 }
-
-const integrationsMappingSchema = {
-  "cal-video": z.literal("integrations:daily"),
-};
-
-const OrganizerAddressSchema = z.object({
-  type: z.literal("inPerson"),
-  address: z.string(),
-  displayLocationPublicly: z.boolean().default(false),
-});
-
-const OrganizerLinkSchema = z.object({
-  type: z.literal("link"),
-  link: z.string().url(),
-  displayLocationPublicly: z.boolean().default(false),
-});
-
-const OrganizerIntegrationSchema = z.object({
-  type: z.union([integrationsMappingSchema["cal-video"], integrationsMappingSchema["cal-video"]]),
-});
-
-const OrganizerPhoneSchema = z.object({
-  type: z.literal("userPhone"),
-  hostPhoneNumber: z.string(),
-  displayLocationPublicly: z.boolean().default(false),
-});
-
-const OrganizerConferencingSchema = z.object({
-  type: z.literal("conferencing"),
-});
-
-const AttendeeAddressSchema = z.object({
-  type: z.literal("attendeeInPerson"),
-});
-
-const AttendeePhoneSchema = z.object({
-  type: z.literal("phone"),
-});
-
-const AttendeeDefinedSchema = z.object({
-  type: z.literal("somewhereElse"),
-});
-
-type OrganizerAddressLocation = z.infer<typeof OrganizerAddressSchema>;
-type OrganizerLinkLocation = z.infer<typeof OrganizerLinkSchema>;
-export type OrganizerIntegrationLocation = z.infer<typeof OrganizerIntegrationSchema>;
-type OrganizerPhoneLocation = z.infer<typeof OrganizerPhoneSchema>;
-type AttendeeAddressLocation = z.infer<typeof AttendeeAddressSchema>;
-type AttendeePhoneLocation = z.infer<typeof AttendeePhoneSchema>;
-type AttendeeDefinedLocation = z.infer<typeof AttendeeDefinedSchema>;
-
-const TransformedLocationSchema = z.union([
-  OrganizerAddressSchema,
-  OrganizerLinkSchema,
-  OrganizerIntegrationSchema,
-  OrganizerPhoneSchema,
-  OrganizerConferencingSchema,
-  AttendeeAddressSchema,
-  AttendeePhoneSchema,
-  AttendeeDefinedSchema,
-]);
-export const TransformedLocationsSchema = z.array(TransformedLocationSchema);
