+++ packages/lib/event-types/transformers/index.ts
@@ -1,2 +1,3 @@
 export * from "./api-to-internal";
 export * from "./internal-to-api";
+export * from "./internal/locations";
